Refactor Constant and Variable classes
- Added namespaces `System.Linq.Expressions` and `System.Numerics` in `Constant.cs`.
- Created the class `Constant<TType>` in `SharpGrad.DifEngine` with methods `GenerateForwardExpression` and an implicit conversion operator.
- Modified the implicit conversion operator in `Value.cs` to return `Constant<TType>` instead of `Variable<TType>`.
- Removed `Constant<TType>` from `Variable.cs` and added an implicit conversion operator from `TType` to `Variable<TType>`.
- Added a static instance counter `InstanceCount` in `Constant<TType>` and `Variable<TType>`.
