+++ SharpGrad/Variable.cs
@@ -11,19 +11,14 @@ namespace SharpGrad.DifEngine
         : Value<TType>(data, name, childs)
         where TType : INumber<TType>
     {
+        private static int InstanceCount = 0;
+
         public override Expression GenerateForwardExpression()
         {
             return Expression.Field(Expression.Constant(this), nameof(Data));
         }
-    }
 
-    public class Constant<TType>(TType data, string name, params Value<TType>[] childs)
-        : Value<TType>(data, name, childs)
-        where TType : INumber<TType>
-    {
-        public override Expression GenerateForwardExpression()
-        {
-            return Expression.Constant(Data);
-        }
+        public static implicit operator Variable<TType>(TType d)
+            => new Variable<TType>(d, $"v{InstanceCount++}");
     }
 }
